% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_poisson.R
\name{r_poisson}
\alias{r_poisson}
\title{Generate random Poisson distributed numbers}
\usage{
r_poisson(n, lambda)
}
\arguments{
\item{n}{Number of random numbers to generate. Must be a positive integer.}

\item{lambda}{The mean (and variance) of the Poisson distribution. Must be positive.}
}
\value{
A numeric vector of n random numbers from Poisson(lambda)
}
\description{
This function generates a user-specified number of random numbers
from a Poisson distribution.
}
\examples{
# Generate 100 random numbers from Poisson(mean=5)
r_poisson(n = 100, lambda = 5)

# Generate 50 random numbers from Poisson(mean=2.5)
r_poisson(n = 50, lambda = 2.5)
}
